SGNewItemNameSpecialCharacterTest class deleted, transfer tests to _NewItemTest Class
