+++ src/test/java/_NewItemTest.java
@@ -8,11 +8,14 @@ import runner.BaseTest;
 
 public class _NewItemTest extends BaseTest {
 
+    private static final By NEW_ITEM = By.className("task-link-text");
+    private static final By INPUT_LINE = By.id("name");
     private final String DESCRIPTION_FIELD = "description";
     private final String GITHUB_URL = "_.projectUrlStr";
     private final String NAME = "NJ";
     private final String DESCRIPTION_INPUT = "New Project created by TA";
     private final String URL_INPUT = "https://github.com/SergeiDemyanenko/JenkinsQA_04/";
+    private static final By ERROR_TEXT = By.id("itemname-invalid");
 
     public void homePage() {
         getDriver().findElement(By.id("jenkins-home-link")).click();
@@ -80,5 +83,27 @@ public class _NewItemTest extends BaseTest {
         asserts.assertEquals(getDriver().findElement(By.name(DESCRIPTION_FIELD)).getText(), DESCRIPTION_INPUT);
         asserts.assertEquals(getDriver().findElement(By.name(GITHUB_URL)).getAttribute("value"), URL_INPUT);
         asserts.assertAll();
+    }
+
+    @Test
+    public void testNameWithOpenSquareBracket() {
+        String expectedResult = "» ‘[’ is an unsafe character";
+
+        getDriver().findElement(NEW_ITEM).click();
+        getDriver().findElement(INPUT_LINE).sendKeys("[");
+        String actualResult = getDriver().findElement(ERROR_TEXT).getText();
+
+        Assert.assertEquals(actualResult, expectedResult);
+    }
+
+    @Test
+    public void testNameWithClosedSquareBracket() {
+        String expectedResult = "» ‘]’ is an unsafe character";
+
+        getDriver().findElement(NEW_ITEM).click();
+        getDriver().findElement(INPUT_LINE).sendKeys("]");
+        String actualResult = getDriver().findElement(ERROR_TEXT).getText();
+
+        Assert.assertEquals(actualResult, expectedResult);
     }
 }
